pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
description = "AI image generation and editing CLI tools for Apple Silicon (MPS)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers[torch]",
    "peft",
    "pillow>=12.1.0",
    "requests",
    "torch>=2.10.0",
    "torchvision",
    "transformers>=4.57.6",
]

[project.entry-points."console_scripts"]
"igm.flux-gen" = "igm.ai_tools.flux_gen:main"
"igm.flux-edit" = "igm.ai_tools.flux_edit:main"
"igm.z-gen" = "igm.ai_tools.z_gen:main"
"igm.qwen-gen" = "igm.ai_tools.qwen_gen:main"
"igm.qwen-desc" = "igm.ai_tools.qwen3vl_desc:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["igm*"]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }
